Add RedisDoFn
